double arithmetic

+++ x-pack/plugin/sql/src/main/java/org/elasticsearch/xpack/sql/action/compute/aggregation/AvgAggregator.java
@@ -44,7 +44,7 @@ class AvgAggregator implements AggregatorFunction {
         Block block = page.getBlock(channel);
         AvgState state = this.state;
         for (int i = 0; i < block.getPositionCount(); i++) {
-            state.add(block.getLong(i));
+            state.add(block.getDouble(i));
         }
         state.count += block.getPositionCount();
     }
